+++ nix/lib/default.nix
@@ -721,13 +721,13 @@ in rec {
               echo "################################################################"
               echo "## flake8 ######################################################"
               echo "################################################################"
-              flake8
+              flake8 -v
               echo "################################################################"
 
               echo "################################################################"
               echo "## pytest ######################################################"
               echo "################################################################"
-              pytest tests/
+              pytest tests/ -vvv -s
               echo "################################################################"
             '';
