+++ nix/requirements_override.nix
@@ -82,7 +82,7 @@ in skipOverrides {
       echo "################################################################"
       echo "## flake8 ######################################################"
       echo "################################################################"
-      flake8
+      flake8 -v
       echo "################################################################"
 
       echo "################################################################"
