+++ src/releng_mapper/releng_mapper/__init__.py
@@ -13,6 +13,7 @@ import releng_mapper.config
 def create_app(config=None):
     app = backend_common.create_app(
         name=releng_mapper.config.PROJECT_NAME,
+        config=config,
         extensions=[
             'log',
             'security',
