+++ src/releng_tooltool/releng_tooltool/__init__.py
@@ -13,6 +13,7 @@ import releng_tooltool.config
 def create_app(config=None):
     app = backend_common.create_app(
         name=releng_tooltool.config.PROJECT_NAME,
+        config=config,
         extensions=[
             'log',
             'security',
