+++ src/shipit_uplift/tests/conftest.py
@@ -5,54 +5,36 @@
 
 from __future__ import absolute_import
 
-import copy
-import pytest
-import pickle
-import os
+import backend_common.testing
 import glob
 import json
-
-import backend_common.testing
-from backend_common.mocks import build_header
+import os
+import pickle
+import pytest
 
 
 FIXTURES_DIR = os.path.join(os.path.dirname(__file__), 'fixtures')
 
 
-@pytest.yield_fixture(scope='session')
+@pytest.fixture(scope='session')
 def app():
     '''
     Load shipit_uplift app in test mode
     '''
     import shipit_uplift
 
-    app_config = copy.deepcopy(backend_common.testing.app_config)
-    app_config.update({
+    config = backend_common.testing.get_app_config({
         'SQLALCHEMY_DATABASE_URI': 'sqlite://',
         'SQLALCHEMY_TRACK_MODIFICATIONS': False,
     })
-    app = shipit_uplift.create_app(app_config)
+    app = shipit_uplift.create_app(config)
 
     with app.app_context():
-        backend_common.testing.clear_app(app)
+        backend_common.testing.configure_app(app)
         yield app
 
 
-@pytest.yield_fixture(scope='session')
-def client(app):
-    '''
-    A Flask test client for shipit_uplift
-    with mockups enabled
-    '''
-    from backend_common import mocks
-
-    # Give test client with mockups
-    with app.test_client() as client:
-        with mocks.apply_mockups():
-            yield client
-
-
-@pytest.yield_fixture(scope='session')
+@pytest.fixture(scope='session')
 def bugs(app):
     '''
     Add an analysis and some bugs
@@ -60,25 +42,26 @@ def bugs(app):
     from shipit_uplift.models import (
         BugAnalysis, BugResult, Contributor, BugContributor
     )
-    from backend_common.db import db
 
     # Add 2 analysis
     analysis = []
     for i in ('Dev', 'Release'):
         a = BugAnalysis(name=i)
         a.version = 1  # dummy version
-        db.session.add(a)
+        app.db.session.add(a)
         analysis.append(a)
 
     # Add bugs with real payload
     paths = glob.glob(os.path.join(FIXTURES_DIR, 'bug_*.json'))
     sorted(paths)
+    bugs = []
     for path in paths:
         payload = json.load(open(path))
         bug = BugResult(bugzilla_id=payload['bug']['id'])
         bug.payload = pickle.dumps(payload, 2)
         analysis[0].bugs.append(bug)
-        db.session.add(bug)
+        app.db.session.add(bug)
+        bugs.append(bug)
 
         # Add creator & link to bug
         user = payload['analysis']['users']['creator']
@@ -90,10 +73,12 @@ def bugs(app):
         contrib.comment_private = 'hidden comment'
         contrib.comment_public = 'Top Contributor'
         link = BugContributor(bug=bug, contributor=contrib, roles='creator')
-        db.session.add(contrib)
-        db.session.add(link)
+        app.db.session.add(contrib)
+        app.db.session.add(link)
 
-    db.session.commit()
+    app.db.session.commit()
+
+    return bugs
 
 
 def hawk_header(scopes):
@@ -105,7 +90,7 @@ def hawk_header(scopes):
     ext_data = {
         'scopes': scopes,
     }
-    return build_header(client_id, ext_data)
+    return backend_common.testing.build_header(client_id, ext_data)
 
 
 @pytest.yield_fixture(scope='session')
